refactor(openai): split Responses-API translation into responses module
src/openai.rs exceeded the repo's 1000-line file cap after the Codex
Responses-API translation was added. Move the Responses-API request/response
translators (OpenAIResponseRequest, response_to_anthropic,
chat_completion_to_responses, anthropic_to_response) into a dedicated
src/responses.rs module, sharing the field-shaping helpers (extract_text,
translate_tools) via pub(crate). No behaviour change.
