+++ src/lib.rs
@@ -19,6 +19,7 @@ pub mod provider_proxy;
 pub mod providers;
 pub mod proxy;
 pub mod refresh;
+pub mod responses;
 pub mod storage;
 pub mod subscription;
 pub mod subscription_proxy;
